+++ packages/ledger-icp/package.json
@@ -40,6 +40,7 @@
   "peerDependencies": {
     "@dfinity/agent": "^0.19.2",
     "@dfinity/candid": "^0.19.2",
+    "@dfinity/nns-proto": "^0.0.9",
     "@dfinity/principal": "^0.19.2",
     "@dfinity/utils": "^0.0.23"
   }
