+++ packages/ledger-icp/src/constants/canister_ids.ts
@@ -1,0 +1,5 @@
+import { Principal } from "@dfinity/principal";
+
+export const MAINNET_LEDGER_CANISTER_ID = Principal.fromText(
+  "ryjl3-tyaaa-aaaaa-aaaba-cai",
+);
