+++ packages/ledger-icp/src/index.ts
@@ -1,1 +1,4 @@
-export * from "./ledger.canister";
+export { AccountIdentifier, SubAccount } from "./account_identifier";
+export * from "./errors/ledger.errors";
+export { LedgerCanister } from "./ledger.canister";
+export * from "./types/ledger.options";
