+++ packages/ledger-icp/src/types/common.ts
@@ -1,0 +1,3 @@
+export type AccountIdentifier = string;
+export type BlockHeight = bigint;
+export type E8s = bigint;
