+++ packages/nns/package.json
@@ -53,6 +53,7 @@
   "peerDependencies": {
     "@dfinity/agent": "^0.19.2",
     "@dfinity/candid": "^0.19.2",
+    "@dfinity/ledger-icp": "^0.0.1",
     "@dfinity/nns-proto": "^0.0.9",
     "@dfinity/principal": "^0.19.2",
     "@dfinity/utils": "^0.0.23"
