+++ packages/nns/src/constants/canister_ids.ts
@@ -8,10 +8,6 @@ export const MAINNET_GOVERNANCE_CANISTER_ID = Principal.fromText(
   "rrkah-fqaaa-aaaaa-aaaaq-cai",
 );
 
-export const MAINNET_LEDGER_CANISTER_ID = Principal.fromText(
-  "ryjl3-tyaaa-aaaaa-aaaba-cai",
-);
-
 export const MAINNET_GENESIS_TOKEN_CANISTER_ID = Principal.fromText(
   "renrk-eyaaa-aaaaa-aaada-cai",
 );
