+++ packages/nns/src/index.ts
@@ -1,18 +1,14 @@
 export type { RewardEvent } from "../candid/governance";
 export type { DeployedSns } from "../candid/sns_wasm";
-export { AccountIdentifier, SubAccount } from "./account_identifier";
 export * from "./enums/governance.enums";
 export * from "./errors/governance.errors";
-export * from "./errors/ledger.errors";
 export { GenesisTokenCanister } from "./genesis_token.canister";
 export { GovernanceCanister } from "./governance.canister";
 export { ICP } from "./icp";
-export { LedgerCanister } from "./ledger.canister";
 export { SnsWasmCanister } from "./sns_wasm.canister";
 export * from "./types/common";
 export * from "./types/governance.options";
 export * from "./types/governance_converters";
-export * from "./types/ledger.options";
 export type { SnsWasmCanisterOptions } from "./types/sns_wasm.options";
 export * from "./utils/account_identifier.utils";
 export * from "./utils/accounts.utils";
